// src/app/core/models/enums/configration-columns.ts
export enum ConfigrationColumns {
	introductionVideoUrl = 2,
    supportEmail = 3,
    mjlsiSystemBeforeMeetingVideoUrl = 4,
    explainCreateMeetingVideoUrl = 5,
    manageBoardMeetingVideoUrl = 6,
    manageBoardMeetingExtraVideoUrl = 7,
    introductionArabicPdfUrl = 8,
    introductionEnglishPdfUrl = 9,
    thirdPdfUrl = 10
}